@@ -3,17 +3,17 @@ apply from: '../properties-util.gradle'
 apply plugin: 'walle'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "com.sunrain.timetablev4"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 29
         versionName "2.2.1"
         resConfigs "zh"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     signingConfigs {
@@ -56,26 +56,22 @@ android {
     }
 }
 
-ext.versions = [
-        'support': '27.1.1'
-]
-
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.0', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
     implementation 'tech.gujin:toast-util:1.1.1'
     implementation 'com.google.zxing:core:3.3.3'
     implementation 'com.tencent.bugly:crashreport:2.6.6.1'
     implementation 'com.meituan.android.walle:library:1.1.6'
-    implementation 'com.android.support:exifinterface:' + versions.support
-    implementation 'com.android.support:support-compat:' + versions.support
-    implementation 'com.android.support:support-v13:' + versions.support
-    implementation 'com.android.support:support-annotations:' + versions.support
-    implementation 'com.android.support:support-fragment:' + versions.support
-    implementation 'com.android.support.constraint:constraint-layout:1.1.1'
+    implementation 'androidx.exifinterface:exifinterface:1.0.0'
+    implementation 'androidx.core:core:1.0.1'
+    implementation 'androidx.legacy:legacy-support-v13:1.0.0'
+    implementation 'androidx.annotation:annotation:1.0.2'
+    implementation 'androidx.fragment:fragment:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     debugImplementation 'com.squareup.leakcanary:leakcanary-android:1.5.4'
     releaseImplementation 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.4'
 }
