@@ -2,7 +2,7 @@ package com.sunrain.timetablev4.adapter.course_management;
 
 import android.content.Context;
 import android.content.DialogInterface;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
