@@ -1,6 +1,6 @@
 package com.sunrain.timetablev4.bean;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public class ClassTimeBean {
 
