@@ -1,7 +1,7 @@
 package com.sunrain.timetablev4.ui.fragment.settings;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
