@@ -1,6 +1,6 @@
 package com.sunrain.timetablev4.utils;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Base64;
 
 import java.io.ByteArrayOutputStream;
